feat: Add Terraform providers and VPC module
Added two new Terraform providers: `google` and `google-beta`. Both are set to version 6.25.0 with specific constraints and hashes for security.

Also, a new VPC module has been introduced in the main.tf file. This includes configurations for three subnets, secondary ranges for two of them, and routing rules. The project ID is set to "lab-courtney" and the network name is "example-vpc".